Ticket #6589: Move methods for project media export to a concern
Fix CC issue: too many lines on class
