+++ app/models/project_media.rb
@@ -5,6 +5,7 @@ class ProjectMedia < ActiveRecord::Base
   include ProjectMediaAssociations
   include ProjectMediaCreators
   include ProjectMediaEmbed
+  include ProjectMediaExport
   include Versioned
   include NotifyEmbedSystem
   include ValidationsHelper
@@ -225,16 +226,6 @@ class ProjectMedia < ActiveRecord::Base
     ability ||= Ability.new
     perms["restore ProjectMedia"] = ability.can?(:restore, self)
     perms
-  end
-
-  def contributing_users_count
-    self.annotations.where(annotator_type: 'User').group_by(&:annotator_id).size
-  end
-
-  def time_to_status(position)
-    statuses_log = self.get_versions_log.where(event_type: 'update_status')
-    return '' if statuses_log.empty? || (statuses_log.size == 1 && position == :first)
-    (statuses_log.send(position).created_at - self.created_at).to_i
   end
 
   private
